Move cli-specific FS logic into its own file
